slight refactor of the routes file
